Add worker to load urls from a file
This worker will read lines from a file, as a stream, will convert them
to a `PageInfo` object and will put them on a queue.

Internally it uses a single thread executor. It's main purpose is to
feed the queue with data from a file.

One could create several UrlFileLoader classes for distinct files and
use them produce data for the same queue.