Add the body requester logic
This service gets urls from a queue, and performs an HTTP request to
obtain the response body. While doing this I felt the need to have some
marker that would represent the end of the work to be done.

The PageInfo.POISON was created for that. It's a special marker object
that notifies the services on the pipeline that the work is completed.

Note that this still need some work, mainly handling timeouts and
redirects. This commit acts as a base for future work on this.