Add technology processor engine
Added a new service to find technology matches on a stream of PageInfos.
It has an internal collection of matchers that is easy to change.

A TechnologyMatcher interface was created to support this logic, that
specific implementations can use.

The TechnologyProcessor is very similar in logic to the BodyRequester,
maybe we can extract a base class in the future.