Delay poison until work done
Before this patch, as soon as a service received the poison, it would
submit it. This happened even if there were still work being done. So
the OutputSink received it and finished and actually didn't log
anything.

Now the main services will wait until all is completed before passing
the poison.