Back pressurize main services
By information given by the metris monitor, the system would load a 1M
entries right way, bypass the queues, and would create 1M taks for the
executor service. Note that everything worked fine, but the system
consumed lots of resources, going from 700Mb to 1Gb ram, and effectly
not streaming, but loading everything as once.

This patch adds back pressure. If the main services don't have workers
available, they will wait and not consume the queues. With this change
the memory is now at the range 20 Mb - 90 Mb.